Improve idle timer and window behaviour
The mode now creates a new window if necessary to display eww in the
same frame, but not on top of the current buffer. It also updates
every second. To turn the preview off again, run markdown-preview-eww
a second time.

There's the initial vestiges of support for more than one preview, but
that's not really done yet. In particular, you'd need to make a
tempfile for the output.